Refactoring cli code initial commit
