+++ src/connectedk8s/azext_connectedk8s/_utils.py
@@ -5,33 +5,28 @@
 
 import os
 import shutil
-import subprocess
-from subprocess import Popen, PIPE
-import time
-import requests
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
-import json
+import platform
+from base64 import b64encode
 
-from knack.util import CLIError
+import yaml
+from requests.adapters import HTTPAdapter
+from Crypto.IO import PEM
+from Crypto.PublicKey import RSA
+from Crypto.Util import asn1
+from psutil import process_iter, NoSuchProcess, AccessDenied, ZombieProcess, net_connections
 from knack.log import get_logger
 from knack.prompting import NoTTYException, prompt_y_n
-from azure.cli.core.commands.client_factory import get_subscription_id
 from azure.cli.core.util import send_raw_request
 from azure.cli.core import telemetry
-from azure.core.exceptions import ResourceNotFoundError
-from msrest.exceptions import AuthenticationError, HttpOperationError, TokenExpiredError
-from msrest.exceptions import ValidationError as MSRestValidationError
-from msrestazure.azure_exceptions import CloudError
-from kubernetes.client.rest import ApiException
-from azext_connectedk8s._client_factory import _resource_client_factory, _resource_providers_client
+from azure.cli.core.azclierror import CLIInternalError, ValidationError, FileOperationError, ClientRequestError
+from azure.cli.core.azclierror import ManualInterrupt
+from azext_connectedk8s._helm_core_utils import HelmCoreUtils
 import azext_connectedk8s._constants as consts
-from kubernetes import client as kube_client
-from azure.cli.core.azclierror import CLIInternalError, ClientRequestError, ArgumentUsageError, ManualInterrupt, AzureResponseError, AzureInternalError, ValidationError
+import azext_connectedk8s._kube_core_utils as kube_core_utils
+from .vendored_sdks.models import ConnectedCluster, ConnectedClusterIdentity
 
 logger = get_logger(__name__)
 
-# pylint: disable=line-too-long
 # pylint: disable=bare-except
 
 
@@ -50,29 +45,12 @@ class TimeoutHTTPAdapter(HTTPAdapter):
         return super().send(request, **kwargs)
 
 
-def validate_location(cmd, location):
-    subscription_id = get_subscription_id(cmd.cli_ctx)
-    rp_locations = []
-    resourceClient = _resource_client_factory(cmd.cli_ctx, subscription_id=subscription_id)
-    try:
-        providerDetails = resourceClient.providers.get('Microsoft.Kubernetes')
-    except Exception as e:  # pylint: disable=broad-except
-        arm_exception_handler(e, consts.Get_ResourceProvider_Fault_Type, 'Failed to fetch resource provider details')
-    for resourceTypes in providerDetails.resource_types:
-        if resourceTypes.resource_type == 'connectedClusters':
-            rp_locations = [location.replace(" ", "").lower() for location in resourceTypes.locations]
-            if location.lower() not in rp_locations:
-                telemetry.set_exception(exception='Location not supported', fault_type=consts.Invalid_Location_Fault_Type,
-                                        summary='Provided location is not supported for creating connected clusters')
-                raise ArgumentUsageError("Connected cluster resource creation is supported only in the following locations: " +
-                                         ', '.join(map(str, rp_locations)), recommendation="Use the --location flag to specify one of these locations.")
-            break
-
-
 def get_chart_path(registry_path, kube_config, kube_context):
     # Pulling helm chart from registry
     os.environ['HELM_EXPERIMENTAL_OCI'] = '1'
-    pull_helm_chart(registry_path, kube_config, kube_context)
+
+    helm_core_utils = HelmCoreUtils(kube_config, kube_context)
+    helm_core_utils.pull_helm_chart(registry_path)
 
     # Exporting helm chart after cleanup
     chart_export_path = os.path.join(os.path.expanduser('~'), '.azure', 'AzureArcCharts')
@@ -80,8 +58,10 @@ def get_chart_path(registry_path, kube_config, kube_context):
         if os.path.isdir(chart_export_path):
             shutil.rmtree(chart_export_path)
     except:
-        logger.warning("Unable to cleanup the azure-arc helm charts already present on the machine. In case of failure, please cleanup the directory '%s' and try again.", chart_export_path)
-    export_helm_chart(registry_path, chart_export_path, kube_config, kube_context)
+        logger.warning("Unable to cleanup the azure-arc helm charts already present on the machine." +
+                       " In case of failure, please cleanup the directory '%s' and try again.",
+                       chart_export_path)
+    helm_core_utils.export_helm_chart(registry_path, chart_export_path)
 
     # Returning helm chart path
     helm_chart_path = os.path.join(chart_export_path, 'azure-arc-k8sagents')
@@ -89,56 +69,14 @@ def get_chart_path(registry_path, kube_config, kube_context):
     return chart_path
 
 
-def pull_helm_chart(registry_path, kube_config, kube_context):
-    cmd_helm_chart_pull = ["helm", "chart", "pull", registry_path]
-    if kube_config:
-        cmd_helm_chart_pull.extend(["--kubeconfig", kube_config])
-    if kube_context:
-        cmd_helm_chart_pull.extend(["--kube-context", kube_context])
-    response_helm_chart_pull = subprocess.Popen(cmd_helm_chart_pull, stdout=PIPE, stderr=PIPE)
-    _, error_helm_chart_pull = response_helm_chart_pull.communicate()
-    if response_helm_chart_pull.returncode != 0:
-        telemetry.set_exception(exception=error_helm_chart_pull.decode("ascii"), fault_type=consts.Pull_HelmChart_Fault_Type,
-                                summary='Unable to pull helm chart from the registry')
-        raise CLIInternalError("Unable to pull helm chart from the registry '{}': ".format(registry_path) + error_helm_chart_pull.decode("ascii"))
-
-
-def export_helm_chart(registry_path, chart_export_path, kube_config, kube_context):
-    cmd_helm_chart_export = ["helm", "chart", "export", registry_path, "--destination", chart_export_path]
-    if kube_config:
-        cmd_helm_chart_export.extend(["--kubeconfig", kube_config])
-    if kube_context:
-        cmd_helm_chart_export.extend(["--kube-context", kube_context])
-    response_helm_chart_export = subprocess.Popen(cmd_helm_chart_export, stdout=PIPE, stderr=PIPE)
-    _, error_helm_chart_export = response_helm_chart_export.communicate()
-    if response_helm_chart_export.returncode != 0:
-        telemetry.set_exception(exception=error_helm_chart_export.decode("ascii"), fault_type=consts.Export_HelmChart_Fault_Type,
-                                summary='Unable to export helm chart from the registry')
-        raise CLIInternalError("Unable to export helm chart from the registry '{}': ".format(registry_path) + error_helm_chart_export.decode("ascii"))
-
-
-def add_helm_repo(kube_config, kube_context):
-    repo_name = os.getenv('HELMREPONAME')
-    repo_url = os.getenv('HELMREPOURL')
-    cmd_helm_repo = ["helm", "repo", "add", repo_name, repo_url]
-    if kube_config:
-        cmd_helm_repo.extend(["--kubeconfig", kube_config])
-    if kube_context:
-        cmd_helm_repo.extend(["--kube-context", kube_context])
-    response_helm_repo = Popen(cmd_helm_repo, stdout=PIPE, stderr=PIPE)
-    _, error_helm_repo = response_helm_repo.communicate()
-    if response_helm_repo.returncode != 0:
-        telemetry.set_exception(exception=error_helm_repo.decode("ascii"), fault_type=consts.Add_HelmRepo_Fault_Type,
-                                summary='Failed to add helm repository')
-        raise CLIInternalError("Unable to add repository {} to helm: ".format(repo_url) + error_helm_repo.decode("ascii"))
-
-
 def get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood=None, release_train_dogfood=None):
     # Setting uri
-    get_chart_location_url = "{}/{}/GetLatestHelmPackagePath?api-version=2019-11-01-preview".format(config_dp_endpoint, 'azure-arc-k8sagents')
+    get_chart_location_url = "{}/{}/GetLatestHelmPackagePath?api-version=2019-11-01-preview" \
+                             .format(config_dp_endpoint, 'azure-arc-k8sagents')
     release_train = os.getenv('RELEASETRAIN') if os.getenv('RELEASETRAIN') else 'stable'
     if dp_endpoint_dogfood:
-        get_chart_location_url = "{}/azure-arc-k8sagents/GetLatestHelmPackagePath?api-version=2019-11-01-preview".format(dp_endpoint_dogfood)
+        get_chart_location_url = "{}/azure-arc-k8sagents/GetLatestHelmPackagePath?api-version=2019-11-01-preview" \
+                                 .format(dp_endpoint_dogfood)
         if release_train_dogfood:
             release_train = release_train_dogfood
     uri_parameters = ["releaseTrain={}".format(release_train)]
@@ -146,7 +84,8 @@ def get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood=None, release
 
     # Sending request
     try:
-        r = send_raw_request(cmd.cli_ctx, 'post', get_chart_location_url, uri_parameters=uri_parameters, resource=resource)
+        r = send_raw_request(cmd.cli_ctx, 'post', get_chart_location_url, uri_parameters=uri_parameters,
+                             resource=resource)
     except Exception as e:
         telemetry.set_exception(exception=e, fault_type=consts.Get_HelmRegistery_Path_Fault_Type,
                                 summary='Error while fetching helm chart registry path')
@@ -159,74 +98,10 @@ def get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood=None, release
                                     summary='Error while fetching helm chart registry path')
             raise CLIInternalError("Error while fetching helm chart registry path from JSON response: " + str(e))
     else:
-        telemetry.set_exception(exception='No content in response', fault_type=consts.Get_HelmRegistery_Path_Fault_Type,
+        telemetry.set_exception(exception='No content in response',
+                                fault_type=consts.Get_HelmRegistery_Path_Fault_Type,
                                 summary='No content in acr path response')
         raise CLIInternalError("No content was found in helm registry path response.")
-
-
-def arm_exception_handler(ex, fault_type, summary, return_if_not_found=False):
-    if isinstance(ex, AuthenticationError):
-        telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-        raise AzureResponseError("Authentication error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-    if isinstance(ex, TokenExpiredError):
-        telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-        raise AzureResponseError("Token expiration error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-    if isinstance(ex, HttpOperationError):
-        status_code = ex.response.status_code
-        if status_code == 404 and return_if_not_found:
-            return
-        if status_code // 100 == 4:
-            telemetry.set_user_fault()
-        telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-        if status_code // 100 == 5:
-            raise AzureInternalError("Http operation error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-        raise AzureResponseError("Http operation error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-    if isinstance(ex, MSRestValidationError):
-        telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-        raise AzureResponseError("Validation error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-    if isinstance(ex, CloudError):
-        status_code = ex.status_code
-        if status_code == 404 and return_if_not_found:
-            return
-        if status_code // 100 == 4:
-            telemetry.set_user_fault()
-        telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-        if status_code // 100 == 5:
-            raise AzureInternalError("Cloud error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-        raise AzureResponseError("Cloud error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-    if isinstance(ex, ResourceNotFoundError) and return_if_not_found:
-        return
-
-    telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-    raise ClientRequestError("Error occured while making ARM request: " + str(ex) + "\nSummary: {}".format(summary))
-
-
-def kubernetes_exception_handler(ex, fault_type, summary, error_message='Error occured while connecting to the kubernetes cluster: ',
-                                 message_for_unauthorized_request='The user does not have required privileges on the kubernetes cluster to deploy Azure Arc enabled Kubernetes agents. Please ensure you have cluster admin privileges on the cluster to onboard.',
-                                 message_for_not_found='The requested kubernetes resource was not found.', raise_error=True):
-    telemetry.set_user_fault()
-    if isinstance(ex, ApiException):
-        status_code = ex.status
-        if status_code == 403:
-            logger.warning(message_for_unauthorized_request)
-        elif status_code == 404:
-            logger.warning(message_for_not_found)
-        else:
-            logger.debug("Kubernetes Exception: " + str(ex))
-        if raise_error:
-            telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-            raise ValidationError(error_message + "\nError Response: " + str(ex.body))
-    else:
-        if raise_error:
-            telemetry.set_exception(exception=ex, fault_type=fault_type, summary=summary)
-            raise ValidationError(error_message + "\nError: " + str(ex))
-        else:
-            logger.debug("Kubernetes Exception: " + str(ex))
 
 
 def validate_infrastructure_type(infra):
@@ -251,102 +126,11 @@ def get_values_file():
     return values_file_provided, values_file
 
 
-def ensure_namespace_cleanup(configuration):
-    api_instance = kube_client.CoreV1Api(kube_client.ApiClient(configuration))
-    timeout = time.time() + 180
-    while True:
-        if time.time() > timeout:
-            telemetry.set_user_fault()
-            logger.warning("Namespace 'azure-arc' still in terminating state. Please ensure that you delete the 'azure-arc' namespace before onboarding the cluster again.")
-            return
-        try:
-            api_response = api_instance.list_namespace(field_selector='metadata.name=azure-arc')
-            if not api_response.items:
-                return
-            time.sleep(5)
-        except Exception as e:  # pylint: disable=broad-except
-            logger.warning("Error while retrieving namespace information: " + str(e))
-            kubernetes_exception_handler(e, consts.Get_Kubernetes_Namespace_Fault_Type, 'Unable to fetch kubernetes namespace',
-                                         raise_error=False)
-
-
-def delete_arc_agents(release_namespace, kube_config, kube_context, configuration):
-    cmd_helm_delete = ["helm", "delete", "azure-arc", "--namespace", release_namespace]
-    if kube_config:
-        cmd_helm_delete.extend(["--kubeconfig", kube_config])
-    if kube_context:
-        cmd_helm_delete.extend(["--kube-context", kube_context])
-    response_helm_delete = Popen(cmd_helm_delete, stdout=PIPE, stderr=PIPE)
-    _, error_helm_delete = response_helm_delete.communicate()
-    if response_helm_delete.returncode != 0:
-        if 'forbidden' in error_helm_delete.decode("ascii") or 'Error: warning: Hook pre-delete' in error_helm_delete.decode("ascii") or 'Error: timed out waiting for the condition' in error_helm_delete.decode("ascii"):
-            telemetry.set_user_fault()
-        telemetry.set_exception(exception=error_helm_delete.decode("ascii"), fault_type=consts.Delete_HelmRelease_Fault_Type,
-                                summary='Unable to delete helm release')
-        raise CLIInternalError("Error occured while cleaning up arc agents. " +
-                               "Helm release deletion failed: " + error_helm_delete.decode("ascii") +
-                               " Please run 'helm delete azure-arc' to ensure that the release is deleted.")
-    ensure_namespace_cleanup(configuration)
-
-
-def helm_install_release(chart_path, subscription_id, kubernetes_distro, kubernetes_infra, resource_group_name, cluster_name,
-                         location, onboarding_tenant_id, http_proxy, https_proxy, no_proxy, proxy_cert, private_key_pem,
-                         kube_config, kube_context, no_wait, values_file_provided, values_file, cloud_name, disable_auto_upgrade, enable_custom_locations, custom_locations_oid, onboarding_timeout="300"):
-    cmd_helm_install = ["helm", "upgrade", "--install", "azure-arc", chart_path,
-                        "--set", "global.subscriptionId={}".format(subscription_id),
-                        "--set", "global.kubernetesDistro={}".format(kubernetes_distro),
-                        "--set", "global.kubernetesInfra={}".format(kubernetes_infra),
-                        "--set", "global.resourceGroupName={}".format(resource_group_name),
-                        "--set", "global.resourceName={}".format(cluster_name),
-                        "--set", "global.location={}".format(location),
-                        "--set", "global.tenantId={}".format(onboarding_tenant_id),
-                        "--set", "global.onboardingPrivateKey={}".format(private_key_pem),
-                        "--set", "systemDefaultValues.spnOnboarding=false",
-                        "--set", "global.azureEnvironment={}".format(cloud_name),
-                        "--set", "systemDefaultValues.clusterconnect-agent.enabled=true",
-                        "--output", "json"]
-    # Add custom-locations related params
-    if enable_custom_locations:
-        cmd_helm_install.extend(["--set", "systemDefaultValues.customLocations.enabled=true"])
-        cmd_helm_install.extend(["--set", "systemDefaultValues.customLocations.oid={}".format(custom_locations_oid)])
-    # To set some other helm parameters through file
-    if values_file_provided:
-        cmd_helm_install.extend(["-f", values_file])
-    if disable_auto_upgrade:
-        cmd_helm_install.extend(["--set", "systemDefaultValues.azureArcAgents.autoUpdate={}".format("false")])
-    if https_proxy:
-        cmd_helm_install.extend(["--set", "global.httpsProxy={}".format(https_proxy)])
-    if http_proxy:
-        cmd_helm_install.extend(["--set", "global.httpProxy={}".format(http_proxy)])
-    if no_proxy:
-        cmd_helm_install.extend(["--set", "global.noProxy={}".format(no_proxy)])
-    if proxy_cert:
-        cmd_helm_install.extend(["--set-file", "global.proxyCert={}".format(proxy_cert)])
-    if https_proxy or http_proxy or no_proxy:
-        cmd_helm_install.extend(["--set", "global.isProxyEnabled={}".format(True)])
-    if kube_config:
-        cmd_helm_install.extend(["--kubeconfig", kube_config])
-    if kube_context:
-        cmd_helm_install.extend(["--kube-context", kube_context])
-    if not no_wait:
-        # Change --timeout format for helm client to understand
-        onboarding_timeout = onboarding_timeout + "s"
-        cmd_helm_install.extend(["--wait", "--timeout", "{}".format(onboarding_timeout)])
-    response_helm_install = Popen(cmd_helm_install, stdout=PIPE, stderr=PIPE)
-    _, error_helm_install = response_helm_install.communicate()
-    if response_helm_install.returncode != 0:
-        if ('forbidden' in error_helm_install.decode("ascii") or 'timed out waiting for the condition' in error_helm_install.decode("ascii")):
-            telemetry.set_user_fault()
-        telemetry.set_exception(exception=error_helm_install.decode("ascii"), fault_type=consts.Install_HelmRelease_Fault_Type,
-                                summary='Unable to install helm release')
-        logger.warning("Please check if the azure-arc namespace was deployed and run 'kubectl get pods -n azure-arc' to check if all the pods are in running state. A possible cause for pods stuck in pending state could be insufficient resources on the kubernetes cluster to onboard to arc.")
-        raise CLIInternalError("Unable to install helm release: " + error_helm_install.decode("ascii"))
-
-
 def flatten(dd, separator='.', prefix=''):
     try:
         if isinstance(dd, dict):
-            return {prefix + separator + k if prefix else k: v for kk, vv in dd.items() for k, v in flatten(vv, separator, kk).items()}
+            return {prefix + separator + k if prefix else k: v for kk, vv in dd.items() \
+                    for k, v in flatten(vv, separator, kk).items()}
         else:
             return {prefix: dd}
     except Exception as e:
@@ -386,48 +170,212 @@ def is_guid(guid):
         return False
 
 
-def try_list_node_fix():
+def check_process(processName):
+    '''
+    Check if there is any running process that contains the given name processName.
+    '''
+    for proc in process_iter():
+        try:
+            if proc.name().startswith(processName):
+                return True
+        except (NoSuchProcess, AccessDenied, ZombieProcess):
+            pass
+    return False
+
+
+def check_if_port_is_open(port):
     try:
-        from kubernetes.client.models.v1_container_image import V1ContainerImage
+        connections = net_connections(kind='inet')
+        for tup in connections:
+            if int(tup[3][1]) == int(port):
+                return True
+    except Exception as e:
+        telemetry.set_exception(exception=e, fault_type=consts.Port_Check_Fault_Type,
+                                summary='Failed to check if port is in use.')
+        if platform.system() != 'Darwin':
+            logger.info("Failed to check if port is in use. " + str(e))
+        return False
+    return False
 
-        def names(self, names):
-            self._names = names
 
-        V1ContainerImage.names = V1ContainerImage.names.setter(names)
-    except Exception as ex:
-        logger.debug("Error while trying to monkey patch the fix for list_node(): {}".format(str(ex)))
+def send_cloud_telemetry(cmd):
+    telemetry.add_extension_event('connectedk8s', {'Context.Default.AzureCLI.AzureCloud': cmd.cli_ctx.cloud.name})
+    cloud_name = cmd.cli_ctx.cloud.name.upper()
+    # Setting cloud name to format that is understood by golang SDK.
+    if cloud_name == consts.PublicCloud_OriginalName:
+        cloud_name = consts.Azure_PublicCloudName
+    elif cloud_name == consts.USGovCloud_OriginalName:
+        cloud_name = consts.Azure_USGovCloudName
+    return cloud_name
 
 
-def check_provider_registrations(cli_ctx):
+def get_config_dp_endpoint(cmd, location):
+    cloud_based_domain = cmd.cli_ctx.cloud.endpoints.active_directory.split('.')[2]
+    config_dp_endpoint = "https://{}.dp.kubernetesconfiguration.azure.{}".format(location, cloud_based_domain)
+    return config_dp_endpoint
+
+
+def get_public_key(key_pair):
+    pubKey = key_pair.publickey()
+    seq = asn1.DerSequence([pubKey.n, pubKey.e])
+    enc = seq.encode()
+    return b64encode(enc).decode('utf-8')
+
+
+def get_private_key(key_pair):
+    privKey_DER = key_pair.exportKey(format='DER')
+    return PEM.encode(privKey_DER, "RSA PRIVATE KEY")
+
+
+def generate_request_payload(location, public_key, tags, kubernetes_distro, kubernetes_infra):
+    # Create connected cluster resource object
+    identity = ConnectedClusterIdentity(
+        type="SystemAssigned"
+    )
+    if tags is None:
+        tags = {}
+    cc = ConnectedCluster(
+        location=location,
+        identity=identity,
+        agent_public_key_certificate=public_key,
+        tags=tags,
+        distribution=kubernetes_distro,
+        infrastructure=kubernetes_infra
+    )
+    return cc
+
+
+def validate_env_file_dogfood(values_file, values_file_provided):
+    if not values_file_provided:
+        telemetry.set_exception(exception='Helm environment file not provided',
+                                fault_type=consts.Helm_Environment_File_Fault_Type,
+                                summary='Helm environment file missing')
+        raise ValidationError("Helm environment file is required when using Dogfood environment" +
+                              " for onboarding the cluster.", recommendation="Please set the" +
+                              " environment variable 'HELMVALUESPATH' to point to the file.")
+
+    with open(values_file, 'r') as f:
+        try:
+            env_dict = yaml.safe_load(f)
+        except Exception as e:
+            telemetry.set_exception(exception=e, fault_type=consts.Helm_Environment_File_Fault_Type,
+                                    summary='Problem loading the helm environment file')
+            raise FileOperationError("Problem loading the helm environment file: " + str(e))
+        try:
+            assert env_dict.get('global').get('azureEnvironment') == 'AZUREDOGFOOD'
+            assert env_dict.get('systemDefaultValues').get('azureArcAgents').get('config_dp_endpoint_override')
+        except Exception as e:
+            telemetry.set_exception(exception=e, fault_type=consts.Helm_Environment_File_Fault_Type,
+                                    summary='Problem loading the helm environment variables')
+            raise FileOperationError("The required helm environment variables for dogfood onboarding are" +
+                                     " either not present in the file or incorrectly set.",
+                                     recommendation="Please check the values 'global.azureEnvironment' " +
+                                     "and 'systemDefaultValues.azureArcAgents.config_dp_endpoint_override'" +
+                                     " in the file.")
+
+    # Return the dp endpoint and release train
+    dp_endpoint = env_dict.get('systemDefaultValues').get('azureArcAgents').get('config_dp_endpoint_override')
+    release_train = env_dict.get('systemDefaultValues').get('azureArcAgents').get('releaseTrain')
+    return dp_endpoint, release_train
+
+def initial_log_warning():
+    logger.warning("Ensure that you have the latest helm version installed before proceeding.")
+    logger.warning("This operation might take a while...\n")
+
+
+def add_telemetry_extenstion_event(connected_cluster, configuration):
+    # Get kubernetes cluster info for telemetry
+    kubernetes_version = kube_core_utils.get_server_version(configuration)
+    if hasattr(connected_cluster, 'distribution') and (connected_cluster.distribution is not None):
+        kubernetes_distro = connected_cluster.distribution
+    else:
+        kubernetes_distro = kube_core_utils.get_kubernetes_distro(configuration)
+
+    if hasattr(connected_cluster, 'infrastructure') and (connected_cluster.infrastructure is not None):
+        kubernetes_infra = connected_cluster.infrastructure
+    else:
+        kubernetes_infra = kube_core_utils.get_kubernetes_infra(configuration)
+
+    add_telemetry_extenstion_event_raw(kubernetes_version, kubernetes_distro, kubernetes_infra)
+
+
+def add_telemetry_extenstion_event_raw(kubernetes_version, kubernetes_distro, kubernetes_infra):
+    kubernetes_properties = {
+        'Context.Default.AzureCLI.KubernetesVersion': kubernetes_version,
+        'Context.Default.AzureCLI.KubernetesDistro': kubernetes_distro,
+        'Context.Default.AzureCLI.KubernetesInfra': kubernetes_infra
+    }
+    telemetry.add_extension_event('connectedk8s', kubernetes_properties)
+
+
+def check_faulty_helm_version(helm_core_utils):
+    # Check helm version
+    helm_version = helm_core_utils.check_helm_version()
+    telemetry.add_extension_event('connectedk8s', {'Context.Default.AzureCLI.HelmVersion': helm_version})
+
+    # Check for faulty pre-release helm versions
+    if "3.3.0-rc" in helm_version:
+        raise ClientRequestError("The current helm version is not supported for azure-arc onboarding.",
+                                 recommendation="Please upgrade helm to a stable version and try again.")
+
+
+def validate_helm_environment(cmd, values_file, values_file_provided):
+    # Validate the helm environment file for Dogfood.
+    dp_endpoint_dogfood = None
+    release_train_dogfood = None
+    azure_cloud = None
+    if cmd.cli_ctx.cloud.endpoints.resource_manager == consts.Dogfood_RMEndpoint:
+        azure_cloud = consts.Azure_DogfoodCloudName
+        dp_endpoint_dogfood, release_train_dogfood = validate_env_file_dogfood(values_file, values_file_provided)
+
+    return azure_cloud, dp_endpoint_dogfood, release_train_dogfood
+
+
+def get_helm_registry_path(cmd, location, helm_core_utils, arc_agent_version, dp_endpoint_dogfood,
+                           release_train_dogfood):
+    # Adding helm repo
+    if os.getenv('HELMREPONAME') and os.getenv('HELMREPOURL'):
+        helm_core_utils.add_helm_repo(os.getenv('HELMREPONAME'), os.getenv('HELMREPOURL'))
+
+    # Setting the config dataplane endpoint
+    config_dp_endpoint = get_config_dp_endpoint(cmd, location)
+
+    # Retrieving Helm chart OCI Artifact location
+    registry_path = os.getenv('HELMREGISTRY') if os.getenv('HELMREGISTRY') \
+        else get_helm_registry(cmd, config_dp_endpoint, dp_endpoint_dogfood, release_train_dogfood)
+
+    reg_path_array = registry_path.split(':')
+    agent_version = reg_path_array[1]
+
+    # Set agent version in registry path
+    if arc_agent_version is not None:
+        agent_version = arc_agent_version
+        registry_path = reg_path_array[0] + ":" + agent_version
+
+    telemetry.add_extension_event('connectedk8s', {'Context.Default.AzureCLI.AgentVersion': agent_version})
+
+    return registry_path
+
+
+def generate_public_private_key():
+    # Generate public-private key pair
     try:
-        rp_client = _resource_providers_client(cli_ctx)
-        cc_registration_state = rp_client.get(consts.Connected_Cluster_Provider_Namespace).registration_state
-        if cc_registration_state != "Registered":
-            telemetry.set_exception(exception="{} provider is not registered".format(consts.Connected_Cluster_Provider_Namespace), fault_type=consts.CC_Provider_Namespace_Not_Registered_Fault_Type,
-                                    summary="{} provider is not registered".format(consts.Connected_Cluster_Provider_Namespace))
-            raise ValidationError("{} provider is not registered. Please register it using 'az provider register -n 'Microsoft.Kubernetes' before running the connect command.".format(consts.Connected_Cluster_Provider_Namespace))
-        kc_registration_state = rp_client.get(consts.Kubernetes_Configuration_Provider_Namespace).registration_state
-        if kc_registration_state != "Registered":
-            telemetry.set_user_fault()
-            logger.warning("{} provider is not registered".format(consts.Kubernetes_Configuration_Provider_Namespace))
-    except ValidationError as e:
-        raise e
-    except Exception as ex:
-        logger.warning("Couldn't check the required provider's registration status. Error: {}".format(str(ex)))
-
-
-def can_create_clusterrolebindings(configuration):
+        key_pair = RSA.generate(4096)
+    except Exception as e:
+        telemetry.set_exception(exception=e, fault_type=consts.KeyPair_Generate_Fault_Type,
+                                summary='Failed to generate public-private key pair')
+        raise CLIInternalError("Failed to generate public-private key pair. " + str(e))
     try:
-        api_instance = kube_client.AuthorizationV1Api(kube_client.ApiClient(configuration))
-        access_review = kube_client.V1SelfSubjectAccessReview(spec={
-            "resourceAttributes": {
-                "verb": "create",
-                "resource": "clusterrolebindings",
-                "group": "rbac.authorization.k8s.io"
-            }
-        })
-        response = api_instance.create_self_subject_access_review(access_review)
-        return response.status.allowed
-    except Exception as ex:
-        logger.warning("Couldn't check for the permission to create clusterrolebindings on this k8s cluster. Error: {}".format(str(ex)))
-        return "Unknown"
+        public_key = get_public_key(key_pair)
+    except Exception as e:
+        telemetry.set_exception(exception=e, fault_type=consts.PublicKey_Export_Fault_Type,
+                                summary='Failed to export public key')
+        raise CLIInternalError("Failed to export public key." + str(e))
+    try:
+        private_key_pem = get_private_key(key_pair)
+    except Exception as e:
+        telemetry.set_exception(exception=e, fault_type=consts.PrivateKey_Export_Fault_Type,
+                                summary='Failed to export private key')
+        raise CLIInternalError("Failed to export private key." + str(e))
+
+    return public_key, private_key_pem
